requirements.txt
streamlit
pandas
mysql-connector-python
python-dotenv
plotly
scikit-learn
imbalanced-learn